+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -11,9 +11,8 @@ import java.util.Arrays;
 import java.util.List;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuProductTestFixture;
-import kitchenpos.menu.domain.MenuProducts;
-import kitchenpos.menu.domain.MenuTestFixture;
+import kitchenpos.menu.testfixture.MenuProductTestFixture;
+import kitchenpos.menu.testfixture.MenuTestFixture;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
@@ -22,7 +21,7 @@ import kitchenpos.menu.validator.MenuValidator;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menugroup.domain.MenuGroupTestFixture;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductTestFixture;
+import kitchenpos.product.testfixture.ProductTestFixture;
 import kitchenpos.product.repository.ProductRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
