+++ src/test/java/kitchenpos/menu/domain/MenuProductTest.java
@@ -3,14 +3,12 @@ package kitchenpos.menu.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.math.BigDecimal;
-import kitchenpos.common.domain.Price;
+import kitchenpos.menu.testfixture.MenuProductTestFixture;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menugroup.domain.MenuGroupTestFixture;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductTestFixture;
+import kitchenpos.product.testfixture.ProductTestFixture;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 
 class MenuProductTest {
