+++ src/test/java/kitchenpos/menu/domain/MenuTest.java
@@ -1,10 +1,12 @@
 package kitchenpos.menu.domain;
 
 import java.math.BigDecimal;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menu.testfixture.MenuProductTestFixture;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menugroup.domain.MenuGroupTestFixture;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductTestFixture;
+import kitchenpos.product.testfixture.ProductTestFixture;
 import org.junit.jupiter.api.BeforeEach;
 
 class MenuTest {
