+++ src/test/java/kitchenpos/menu/testfixture/MenuTestFixture.java
@@ -1,7 +1,9 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.testfixture;
 
 import java.math.BigDecimal;
 import java.util.List;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menugroup.domain.MenuGroup;
 
 public class MenuTestFixture {
