+++ src/test/java/kitchenpos/order/domain/OrderTest.java
@@ -7,16 +7,16 @@ import java.math.BigDecimal;
 import java.util.Arrays;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuProductTestFixture;
-import kitchenpos.menu.domain.MenuProducts;
-import kitchenpos.menu.domain.MenuTestFixture;
+import kitchenpos.menu.testfixture.MenuProductTestFixture;
+import kitchenpos.menu.testfixture.MenuTestFixture;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menugroup.domain.MenuGroupTestFixture;
 import kitchenpos.order.dto.OrderLineItemRequest;
+import kitchenpos.order.testfixture.OrderMenuTestFixture;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableTestFixture;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductTestFixture;
+import kitchenpos.product.testfixture.ProductTestFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
