+++ src/test/java/kitchenpos/order/testfixture/OrderTestFixture.java
@@ -1,6 +1,8 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.testfixture;
 
 import java.util.List;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.ordertable.domain.OrderTable;
 
 public class OrderTestFixture {
