+++ src/test/java/kitchenpos/order/validator/OrderValidatorTest.java
@@ -9,22 +9,22 @@ import java.util.Collections;
 import java.util.Optional;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuProductTestFixture;
-import kitchenpos.menu.domain.MenuTestFixture;
+import kitchenpos.menu.testfixture.MenuProductTestFixture;
+import kitchenpos.menu.testfixture.MenuTestFixture;
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menugroup.domain.MenuGroupTestFixture;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderMenu;
-import kitchenpos.order.domain.OrderMenuTestFixture;
+import kitchenpos.order.testfixture.OrderMenuTestFixture;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTestFixture;
+import kitchenpos.order.testfixture.OrderTestFixture;
 import kitchenpos.order.dto.OrderLineItemRequest;
 import kitchenpos.order.dto.OrderRequest;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableTestFixture;
 import kitchenpos.ordertable.repository.OrderTableRepository;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductTestFixture;
+import kitchenpos.product.testfixture.ProductTestFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
